requirements.txt
canopen
python-can
